moving contents of WebServer.cpp and WebServer.h back together in one file: fixes issue with #define statements (serial debugging in particular)
